Changes to be committed: 	modified:   lib/display/display.h 	modified:   lib/display/splash.h 	- defs now sit correctly inside the block 	new file:   lib/hwinit/hwinit.c 	new file:   lib/hwinit/hwinit.h 	- moves the hardware/periph inits not already part of display to the lib 	modified:   main.c 	- Now relies on hwinit.h
