+++ lib/display/splash.h
@@ -8,13 +8,6 @@
 #ifndef LIB_DISPLAY_SPLASH_H_
 #define LIB_DISPLAY_SPLASH_H_
 
-
-
-
-
-#endif /* LIB_DISPLAY_SPLASH_H_ */
-
-
 #pragma PERSISTENT(splash_bitmap)
 unsigned char splash_bitmap[] = {
      // 'ARCANA_LABS', 128x128px
@@ -147,3 +140,5 @@ unsigned char splash_bitmap[] = {
  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
  0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff
 };
+
+#endif /* LIB_DISPLAY_SPLASH_H_ */
