+++ lib/hwinit/hwinit.h
@@ -1,0 +1,17 @@
+/*
+ * hwinit.h
+ *
+ *  Created on: May 10, 2021
+ *      Author: patches
+ */
+
+#ifndef LIB_HWINIT_HWINIT_H_
+#define LIB_HWINIT_HWINIT_H_
+
+void Init_GPIO(void);
+void Init_Timers(void);
+void Init_RTC(void);
+void Init_Watchdog(void);
+
+
+#endif /* LIB_HWINIT_HWINIT_H_ */
