+++ main.c
@@ -17,9 +17,10 @@
 
 #include "driverlib.h"
 #include "lib/display/display.h"
+#include "lib/hwinit/hwinit.h"
 #include <msp430.h>
 
-#define VERSION " POC-2021-05-08"       //Max 16 characters.
+#define VERSION " POC-2021-05-10"       //Max 16 characters.
 #define SPLASH_DELAY 1000000            //Number of cycles to make the system idle at the splash page idle.
 
 volatile unsigned char VCOM;            // current VCOM state
@@ -97,102 +98,6 @@ void Update_Buttons_Bar(void){
     }
 }
 
-//We need a simple function to initialize some GPIO pins for driving
-//button input as well as the status LEDs
-void Init_GPIO(void) {
-    //We have a couple GPIO pins that most make sense to set up individually.
-    // Some are explicitly or implicitly set up during Init_SPI.
-    GPIO_setAsOutputPin(GPIO_PORT_P1, GPIO_PIN0); // P1.0 is an LED for indicating VCOM state.
-    GPIO_setOutputLowOnPin(GPIO_PORT_P1, GPIO_PIN0); // Initial VCOM = 0
-    GPIO_setAsOutputPin(GPIO_PORT_P1, GPIO_PIN1); // P1.1 is a debugging LED; currently toggles with each keypress.
-    GPIO_setOutputLowOnPin(GPIO_PORT_P1, GPIO_PIN1);
-
-    // Set up debugging panel GPIO
-    #ifdef DEBUG
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN0); //Current use: Unused
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN0);
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN1); //Current use:
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN1);
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN2); //Current use:
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN2);
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN3); //Current use:
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN3);
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN4); //Current use:
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN4);
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN5); //Current use:
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN5);
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN6); //Current use:
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN6);
-    GPIO_setAsOutputPin(GPIO_PORT_P3, GPIO_PIN7); //Current use:
-    GPIO_setOutputLowOnPin(GPIO_PORT_P3, GPIO_PIN7);
-    #endif
-
-    //Inputs Block
-    //This sets up the buttons as indicated below.
-    GPIO_setAsInputPinWithPullUpResistor(GPIO_PORT_P6, GPIO_PIN0); //A = P6.0
-    GPIO_setAsInputPinWithPullUpResistor(GPIO_PORT_P7, GPIO_PIN1); //B = P7.1
-    GPIO_setAsInputPinWithPullUpResistor(GPIO_PORT_P5, GPIO_PIN7); //C = P5.7
-    GPIO_setAsInputPinWithPullUpResistor(GPIO_PORT_P8, GPIO_PIN3); //D = P8.3
-    GPIO_enableInterrupt(GPIO_PORT_P6, GPIO_PIN0);
-    GPIO_enableInterrupt(GPIO_PORT_P7, GPIO_PIN1);
-    GPIO_enableInterrupt(GPIO_PORT_P5, GPIO_PIN7);
-    GPIO_enableInterrupt(GPIO_PORT_P8, GPIO_PIN3);
-    GPIO_selectInterruptEdge(GPIO_PORT_P6,GPIO_PIN0, GPIO_HIGH_TO_LOW_TRANSITION);
-    GPIO_selectInterruptEdge(GPIO_PORT_P7,GPIO_PIN1, GPIO_HIGH_TO_LOW_TRANSITION);
-    GPIO_selectInterruptEdge(GPIO_PORT_P5,GPIO_PIN7, GPIO_HIGH_TO_LOW_TRANSITION);
-    GPIO_selectInterruptEdge(GPIO_PORT_P8,GPIO_PIN3, GPIO_HIGH_TO_LOW_TRANSITION);
-    GPIO_clearInterrupt(GPIO_PORT_P6, GPIO_PIN0);
-    GPIO_clearInterrupt(GPIO_PORT_P7, GPIO_PIN1);
-    GPIO_clearInterrupt(GPIO_PORT_P5, GPIO_PIN7);
-    GPIO_clearInterrupt(GPIO_PORT_P8, GPIO_PIN3);
-}
-
-void Init_Timers(void) {
-    // There are some very important global timers that we need
-    // Currently we only initialize Timer_A which will raise a 1khz interrupt
-    // Allowing us to tick off the milliseconds.
-    // In future this will likely be altered for power mnagement and to use RTC for main
-    // Time tracking.
-    VCOM = MLCD_VCOM;                                   // Starting value setup.
-
-    // Set master clock to 1MHz; at this rate TimerA would interrupt every millisecond.
-    // This is not likely especially power performant.
-    // sets the properties to init Timer_A
-    CS_setDCOFreq(CS_DCORSEL_0, CS_DCOFSEL_0); // Set DCO frequency 1 MHz (? - experimentation suggests this is not true)
-    CS_initClockSignal(CS_SMCLK, CS_DCOCLK_SELECT, CS_CLOCK_DIVIDER_1); //SMCLK = 1 Mhz
-    Timer_A_initUpModeParam initContParam = {0};
-    initContParam.clockSource = TIMER_A_CLOCKSOURCE_SMCLK;
-    initContParam.clockSourceDivider = TIMER_A_CLOCKSOURCE_DIVIDER_16; // Should yield a freq where this counts at roughly 65kHz
-    initContParam.timerInterruptEnable_TAIE = TIMER_A_TAIE_INTERRUPT_DISABLE;
-    initContParam.timerPeriod = 0xFFFF; // Which SHOULD trigger the interrupt 1/second.
-    initContParam.timerClear = TIMER_A_DO_CLEAR;
-    initContParam.captureCompareInterruptEnable_CCR0_CCIE = TIMER_A_CCIE_CCR0_INTERRUPT_DISABLE;
-    initContParam.startTimer = false;
-    Timer_A_initUpMode(TIMER_A0_BASE, &initContParam); // init TA in compare mode for TA2
-    Timer_A_clearTimerInterrupt(TIMER_A0_BASE);
-    Timer_A_setCompareValue(TIMER_A0_BASE, TIMER_A_CAPTURECOMPARE_REGISTER_0, 0xFFFF); // THIS ALSO SETS THE INTERRUPT FREQUENCY!
-    Timer_A_clearCaptureCompareInterrupt(TIMER_A0_BASE, TIMER_A_CAPTURECOMPARE_REGISTER_0);
-    Timer_A_enableInterrupt(TIMER_A0_BASE);
-    Timer_A_enableCaptureCompareInterrupt(TIMER_A0_BASE, TIMER_A_CAPTURECOMPARE_REGISTER_0);
-    // start the timer
-    Timer_A_startCounter(TIMER_A0_BASE, TIMER_A_UP_MODE);
-}
-
-void Init_RTC(void){
-    //We need a default calendar to use as a local epoch. The date just happens to be when we were setting up this function.
-    Calendar defaultTime;
-    defaultTime.Seconds = 0x00;
-    defaultTime.Minutes = 0x00;
-    defaultTime.Hours = 0x00;
-    defaultTime.DayOfWeek = 0x06;  // No convention is specified, so here we treat sunday as 0.
-    defaultTime.DayOfMonth = 0x08;
-    defaultTime.Month = 0x05;
-    defaultTime.Year = 0x2021; // Accepts values up to 4096 (lel) so no unix epoch issue here.
-
-    RTC_C_initCalendar(RTC_C_BASE, &defaultTime, RTC_C_FORMAT_BCD); // Hey struct, get into your registers.
-    RTC_C_startClock(RTC_C_BASE); // Okay let's go!
-}
-
 
 void Init_Buttons(void){
     buttons_state = 0;
@@ -238,20 +143,9 @@ void Print_CurrentTime(void){
 }
 
 int main(void) {
-    // Kosmo: could any of this init go into an init function?
-
-    // Disable the watchdog timer. We might rely on this later, but not for now.
-    WDTCTL = WDTPW | WDTHOLD;
-
-    // Clear P1 IFGs, more as a formality than anything.
-    P1IFG = 0;
-
-    // Clear the OFIFG because occasionally strange IFGs get set that we aren't handling.
-    SFRIFG1 &= ~OFIFG;
-
-    // Without this output pins can be stuck at current state causing apparent freezes.
+    Init_Watchdog();
     PMM_unlockLPM5();
-
+    VCOM = MLCD_VCOM; // Set the initial state of the VCOM bit.
     Init_GPIO();
     Init_Timers();
     Init_RTC();
